❓ chore: fazendo as regras de engocios de tarefa, turma e problema
além de adicionar as regras de negocios, também foram adicionados novos endoints que persistiam os
problemas de uma tarefa e os usuarios de uma turma

+++ src/main/java/com/tcc/joaomyrlla/appcode2know/serviceImpl/ProblemaServiceImpl.java
@@ -1,12 +1,11 @@
 package com.tcc.joaomyrlla.appcode2know.serviceImpl;
 
 import com.tcc.joaomyrlla.appcode2know.dto.ProblemaDTO;
-import com.tcc.joaomyrlla.appcode2know.dto.UsuarioDTO;
 import com.tcc.joaomyrlla.appcode2know.model.Problema;
 import com.tcc.joaomyrlla.appcode2know.model.Usuario;
 import com.tcc.joaomyrlla.appcode2know.repository.ProblemaRepository;
+import com.tcc.joaomyrlla.appcode2know.repository.UsuarioRepository;
 import com.tcc.joaomyrlla.appcode2know.service.IProblemaService;
-import com.tcc.joaomyrlla.appcode2know.service.IUsuarioService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,7 +20,7 @@ public class ProblemaServiceImpl implements IProblemaService {
     ProblemaRepository problemaRepository;
 
     @Autowired
-    IUsuarioService usuarioService;
+    UsuarioRepository usuarioRepository;
 
     public ProblemaServiceImpl(ProblemaRepository problemaRepository) {
         this.problemaRepository = problemaRepository;
@@ -55,8 +54,20 @@ public class ProblemaServiceImpl implements IProblemaService {
     }
 
     public ProblemaDTO add(ProblemaDTO problema) {
+        Optional<Usuario> usuarioOptional = usuarioRepository.findById(problema.getCriadorId());
+
+        if (usuarioOptional.isEmpty()) {
+            throw new RuntimeException("O usuário de Id" + problema.getCriadorId() + " não existe");
+        }
+
+        Usuario usuario = usuarioOptional.get();
+        if (!usuario.isEhProfessor()) {
+            throw new RuntimeException("O usuário de Id" + problema.getCriadorId() + " não é professor");
+        }
+
         Problema novoProblema = new Problema() ;
         BeanUtils.copyProperties(problema, novoProblema);
+        novoProblema.setCriador(usuario);
 
         problemaRepository.save(novoProblema);
         problema.setId(novoProblema.getId());
@@ -66,30 +77,52 @@ public class ProblemaServiceImpl implements IProblemaService {
 
     @Override
     public void delete(Long id, Long usuarioId) {
-//         TODO: Verificar lista de usuarios em busca do usuario com id informado
-//
-//         TODO: Verificar se o usuario buscado tem o atribuito eh_professor como true
-//
-//         TODO: Caso sim deletar
-//
-//         TODO: Caso nao entregar uma excecao
-        problemaRepository.deleteById(id);
+            Optional<Usuario> usuario = usuarioRepository.findById(usuarioId);
+            Optional<Problema> problema = problemaRepository.findById(id);
+            if (usuario.isEmpty()){
+                throw new RuntimeException("O usuário de Id" + usuarioId + " não existe");
+            }
+
+            if(problema.isEmpty()){
+                throw new RuntimeException("Problema com o ID informado não existe");
+            }
+
+            if (!(usuario.get().isEhProfessor()) || !(problema.get().getCriador().getId().equals(usuarioId))) {
+                    throw new RuntimeException("O usuário não tem permissão para deletar a tarefa");
+            }
+
+
+
+            if(!(problema.get().getTarefas().isEmpty())){
+                problema.get().getTarefas().forEach(tarefa -> {
+                    tarefa.getProblemas().remove(problema.get());
+                });
+            }
+            problemaRepository.deleteById(id);
+
     }
 
     @Override
     public ProblemaDTO edit(ProblemaDTO problema, Long usuarioId) {
-//         TODO: Verificar lista de usuarios em busca do usuario com id informado
-//
-//         TODO: Verificar se o usuario buscado tem o atribuito eh_professor como true
-//
-//         TODO: Caso sim editar
-//
-//         TODO: Caso nao entregar uma excecao
+
+        Optional<Usuario> usuario = usuarioRepository.findById(usuarioId);
+        Optional<Problema> problemaOptional = problemaRepository.findById(problema.getId());
+
+        if (usuario.isEmpty()){
+            throw new RuntimeException("O usuário de Id" + usuarioId + " não existe");
+        }
+
+        if(problemaOptional.isEmpty()){
+            throw new RuntimeException("Problema com o ID informado não existe");
+        }
+
+        if (!(usuario.get().isEhProfessor()) || !(problemaOptional.get().getCriador().getId().equals(usuarioId))) {
+            throw new RuntimeException("O usuário não tem permissão para deletar a tarefa");
+        }
+
         Problema problemaEditado = new Problema();
         BeanUtils.copyProperties(problema, problemaEditado);
-
-        Usuario criador = new Usuario();
-        criador.setId(problema.getCriadorId());
+        problemaEditado.setCriador(usuario.get());
 
         problemaRepository.save(problemaEditado);
